refactored board name / edit board name into a new component called ShowOrEditBoard
